make ready for Windows installer and reduced dependency fix

See merge request sdurobotics/RobWork!351

+++ RobWork/ext/rwpqp/PQP/PQP_Compile.h
@@ -62,7 +62,7 @@ inline float fabs(float x) { return (float)fabs((double)x); }*/
 //
 //-------------------------------------------------------------------------
 
-typedef float PQP_REAL;
+typedef double PQP_REAL;
 
 //-------------------------------------------------------------------------
 //
